临时版本 2018062601
1、修改叠加的颜色
2、增加叠加位置计算功能函数
3、删除车道信息获取接口，改从结果附加信息中获取

+++ PLATERECOGNIZE/PLATERECOGNIZE/Camera/BaseCamera.cpp
@@ -1015,7 +1015,7 @@ void BaseCamera::CompressImg(CameraIMG& camImg, DWORD requireSize)
         return;
     }
 
-    DWORD iImgSize = MAX_IMG_SIZE;
+    size_t iImgSize = MAX_IMG_SIZE;
     int iCompressQuality = 80;
     PBYTE pDestImg = new BYTE[iImgSize];
     do
